Adding improvements
- Adding Equals and HashCode methods implementations, there are two
reasons why these implementations are being added:
    - It is one of the best java practices, since these methods are
important to make sure of the uniqueness of objects.
    - related to the previous, adding this is beneficial for the
application since help us to seamless de-duplicate the list of movies
returned in method searchAllMovies from class MoviesSearchServiceImpl.
e.g.: Though Woody Allen appears on 8 records in the list of movies the
movie "Midnight in Paris" appears twice so we this approach we avoid
count it twice since thanks to Equals and HashCode implementations and
the LinkedHashSet we are using the de-duplication of the movie is assured.

- Creating AbstractSearchResults as base class for MoviesSearchResults
class since this open the posibility of different searchResults for
other kind of the objects different than movies that shared the same
contract as movies.